+++ src/main/kotlin/io/sip3/captain/ce/capturing/PcapEngine.kt
@@ -18,8 +18,8 @@ package io.sip3.captain.ce.capturing
 
 import io.micrometer.core.instrument.Metrics
 import io.netty.buffer.Unpooled
-import io.sip3.captain.ce.domain.ByteBufPayload
 import io.sip3.captain.ce.domain.Packet
+import io.sip3.captain.ce.domain.payload.ByteBufPayload
 import io.sip3.captain.ce.pipeline.EthernetHandler
 import io.vertx.core.AbstractVerticle
 import mu.KotlinLogging
