+++ src/main/kotlin/io/sip3/captain/ce/domain/Packet.kt
@@ -16,6 +16,7 @@
 
 package io.sip3.captain.ce.domain
 
+import io.sip3.captain.ce.domain.payload.Payload
 import java.sql.Timestamp
 
 class Packet {
