+++ src/main/kotlin/io/sip3/captain/ce/domain/payload/ByteArrayPayload.kt
@@ -14,11 +14,14 @@
  * limitations under the License.
  */
 
-package io.sip3.captain.ce.domain
+package io.sip3.captain.ce.domain.payload
 
 import io.netty.buffer.ByteBuf
+import io.netty.buffer.Unpooled
 
-interface Payload {
+inline class ByteArrayPayload(val bytes: ByteArray) : Payload {
 
-    fun encode(): ByteBuf
+    override fun encode(): ByteBuf {
+        return Unpooled.wrappedBuffer(bytes)
+    }
 }
