+++ src/main/kotlin/io/sip3/captain/ce/pipeline/RtpHandler.kt
@@ -19,7 +19,7 @@ package io.sip3.captain.ce.pipeline
 import io.sip3.captain.ce.Routes
 import io.sip3.captain.ce.USE_LOCAL_CODEC
 import io.sip3.captain.ce.domain.Packet
-import io.sip3.captain.ce.domain.RtpHeaderPayload
+import io.sip3.captain.ce.domain.payload.RtpHeaderPayload
 import io.vertx.core.Vertx
 import kotlin.experimental.and
