+++ src/main/kotlin/io/sip3/captain/ce/rtcp/RtcpCollector.kt
@@ -19,6 +19,8 @@ package io.sip3.captain.ce.rtcp
 import io.sip3.captain.ce.Routes
 import io.sip3.captain.ce.USE_LOCAL_CODEC
 import io.sip3.captain.ce.domain.*
+import io.sip3.captain.ce.domain.payload.ByteArrayPayload
+import io.sip3.captain.ce.domain.payload.RtpReportPayload
 import io.vertx.core.AbstractVerticle
 import mu.KotlinLogging
 import kotlin.experimental.and
@@ -261,6 +263,7 @@ class RtcpCollector : AbstractVerticle() {
                 rFactor = (R0 - ieEff - iDelay).toFloat()
 
                 // MoS
+                // TODO: Change to `when`
                 if (rFactor < 0) {
                     mos = MOS_MIN
                 } else if (rFactor > 100F) {
